+++ WebApi/Models/ProductExpense.cs
@@ -7,8 +7,8 @@ namespace WebApi.Models
     public class ProductExpense
     {
         [Key]
-        public long ProductExpenseID { get; set; }
-        public long ProductID { get; set; }
+        public int ProductExpenseID { get; set; }
+        public int ProductID { get; set; }
         public decimal Expense { get; set; }
         public int Quantity { get; set; }
         public DateTime TimeStamp { get; set; }
